Move archive handling from event listener to plugin
